refactor: update test server
